MNT: move legacy import hooks to separate module
Python 2 does not like hook objects to be in a replaced module.
