+++ src/diffpy/structure/_legacy_importer.py
@@ -1,0 +1,88 @@
+#!/usr/bin/env python3
+##############################################################################
+#
+# diffpy.structure  Complex Modeling Initiative
+#                   (c) 2017 Brookhaven Science Associates,
+#                   Brookhaven National Laboratory.
+#                   All rights reserved.
+#
+# File coded by:    Pavol Juhas
+#
+# See AUTHORS.txt for a list of people who contributed.
+# See LICENSE.txt for license information.
+#
+##############################################################################
+
+"""
+Support import of old camel-case module names with DeprecationWarning.
+
+The imported camel-case modules are aliases for the current module
+instances.  Their `__name__` attributes are thus all in lower-case.
+
+This module is deprecated and will be removed in the future.
+"""
+
+
+import sys
+import importlib.abc
+from warnings import warn
+
+WMSG = "Module {!r} is deprecated.  Use {!r} instead."
+
+# ----------------------------------------------------------------------------
+
+class FindRenamedStructureModule(importlib.abc.MetaPathFinder):
+
+    prefix = 'diffpy.Structure.'
+
+    @classmethod
+    def find_spec(cls, fullname, path=None, target=None):
+        # only handle submodules of diffpy.Structure
+        if not fullname.startswith(cls.prefix):
+            return None
+        lcname = fullname.lower()
+        spec = importlib.util.find_spec(lcname)
+        if spec is not None:
+            spec.name = fullname
+            spec.loader = MapRenamedStructureModule
+        return spec
+
+# end of class FindRenamedStructureModule
+
+# ----------------------------------------------------------------------------
+
+class MapRenamedStructureModule(importlib.abc.Loader):
+    """
+    Loader for old camel-case module names.
+
+    Import the current module and alias it under the old name.
+    """
+
+    @staticmethod
+    def create_module(spec):
+        lcname = spec.name.lower()
+        mod = importlib.import_module(lcname)
+        sys.modules[spec.name] = mod
+        warn(WMSG.format(spec.name, lcname), DeprecationWarning, stacklevel=2)
+        return mod
+
+
+    @staticmethod
+    def exec_module(module):
+        return
+
+# end of class MapRenamedStructureModule
+
+# ----------------------------------------------------------------------------
+
+# import diffpy.Structure with deprecation warning
+import diffpy.structure
+sys.modules['diffpy.Structure'] = diffpy.structure
+
+warn(WMSG.format('diffpy.Structure', 'diffpy.structure'),
+     DeprecationWarning, stacklevel=2)
+
+# install meta path finder for diffpy.Structure submodules
+sys.meta_path.append(FindRenamedStructureModule)
+
+# End of file
